modified folder structure, other components and added next auth provider
