+++ README.md
@@ -1,4 +1,4 @@
-This is a [Next.js](https://nextjs.org) project bootstrapped with [`create-next-app`](https://github.com/vercel/next.js/tree/canary/packages/create-next-app).
+# Ezi Drop (Smart courier and delivery web application)
 
 ## Getting Started
 
@@ -34,12 +34,4 @@ You can check out [the Next.js GitHub repository](https://github.com/vercel/next
 The easiest way to deploy your Next.js app is to use the [Vercel Platform](https://vercel.com/new?utm_medium=default-template&filter=next.js&utm_source=create-next-app&utm_campaign=create-next-app-readme) from the creators of Next.js.
 
 Check out our [Next.js deployment documentation](https://nextjs.org/docs/app/building-your-application/deploying) for more details.
-let's code your carrer.
-let's code your carrer.
-let's code your carrer.
-let's code your carrer.
-let's code your carrer.
-let's code your carrer.
-let's code your carrer.
-let's code your carrer.
-"True"
+
