+++ package.json
@@ -14,6 +14,7 @@
     "framer-motion": "^12.23.12",
     "lucide-react": "^0.544.0",
     "next": "15.5.3",
+    "next-auth": "^4.24.11",
     "react": "19.1.0",
     "react-dom": "19.1.0",
     "react-hook-form": "^7.62.0",
