+++ src/app/page.js
@@ -1,8 +1,8 @@
-import HowItWorks from "@/components/feature/HowItWorks";
+import HowItWorks from "@/components/homePageComponents/feature/HowItWorks";
 import BannerSection from "@/components/homePageComponents/BannerSection";
 import Faq from "@/components/homePageComponents/faqSection/Faq";
 import NewsLetter from "@/components/homePageComponents/newsletter/NewsLetter";
-import Overview from "@/components/homePageComponents/newsletter/Overview";
+import Overview from "@/components/homePageComponents/Overview";
 import Review from "@/components/homePageComponents/reviewSection/Review";
 
 export default function Home() {
