+++ src/components/sharedComponents/navbar/Navbar.jsx
@@ -3,8 +3,12 @@ import Link from "next/link";
 import React, { useEffect, useState } from "react";
 import { CiMenuBurger } from "react-icons/ci";
 import Sidebar from "./Sidebar";
+import { useSession } from "next-auth/react";
 
 export default function Navbar() {
+
+  const { data: session, status } = useSession();
+
   const [dark, setDark] = useState(false);
   const [openMenu, setOpenMenu] = useState(false);
 
@@ -34,16 +38,17 @@ export default function Navbar() {
   const navLinks = (
     <>
       <Link href={'/'}>Home</Link>
-      <Link href={"/sendpercel"}>Send Parcel</Link>
+      <Link href={"/send-parcel"}>Send Parcel</Link>
       <Link href={"dashboard"}>Dashboard</Link>
       <Link href={"/about"}>About</Link>
+      <Link href={"/contact"}>Contact</Link>
     </>
   );
 
   return (
     <div className="bg-[var(--color-bg)] dark:bg-[var(--color-bg-dark)] px-6 md:px-8 relative">
       <div className="max-w-[1440px] mx-auto flex justify-between items-center h-[100px]">
-        <h1 className="text-2xl font-bold">Ezi Drop </h1>
+        <h1 className="text-2xl font-bold"><Link href={"/"}>Ezi Drop </Link></h1>
         <div className="flex items-center gap-8">
           <nav>
             <ul className="md:flex gap-8 hidden">
@@ -51,9 +56,14 @@ export default function Navbar() {
               {navLinks}
             </ul>
           </nav>
-          <button className="hidden sm:block btn bg-[var(--color-primary)] dark:bg-[var(--color-primary-dark)] rounded-sm text-white border-none">
-            Login
-          </button>
+              <div className='flex gap-4'>
+      {
+        status === "authenticated" ?
+        <button onClick={() => signOut()} className="hidden sm:block btn bg-[var(--color-secondary)] dark:bg-[var(--color-secondary-dark)] rounded-sm text-white border-none">Logout</button>
+       :
+      <Link href={'/login'}><button className="hidden sm:block btn bg-[var(--color-primary)] dark:bg-[var(--color-primary-dark)] rounded-sm text-white border-none">Login</button></Link>
+      }
+    </div>
           <button className="hidden md:block" onClick={toggleTheme}>
             {dark ? "☀️ Light" : "🌙 Dark"}
           </button>{" "}
@@ -63,8 +73,8 @@ export default function Navbar() {
         </div>
       </div>
       {openMenu && (
-        <div className="md:hidden absolute right-0 top-[100px]">
-          <Sidebar navLinks={navLinks} />
+        <div className="md:hidden absolute right-0 top-[100px] z-1000">
+          <Sidebar navLinks={navLinks} status={status} />
         </div>
       )}
     </div>
